Two-step scrape process
Switch to a two-step scraping process:

First we (re-)scrape the homepage (at most once per day), to get a list of
all the individual pages we will need to scrape. Anyone we haven't seen
before gets written to the data, but without any of the data we'll get
from the individual pages.

Secondly, we get a list of pages we still need to scrape, by looking for
all the people who don't yet have a 'term' field filled. (Later we can
extend this to have a `last_scraped` field too, and re-fetch them again
after a certain interval), and iterate over those.

We also trap any errors (e.g. pages unable to load), warn about the
error, sleep for a little while longer, and then continue again, rather
than terminating on error.